separate examples.py router module with get/post to replace `/examples` and `/validate` endpoints
